[JSX] Added a LinkElement to jsx library (#3347)
This creates a react form element that has an href hyperlink destination which fits takes up a row in the LORIS bootstrap theme.

This is accomplished by creating a new form element in the loris/jsx library that is modelled after the "Static Element". However, the "Link Element" also includes an 'href' prop type which allows the developer to set a hyperlink destination for the given element.